+++ MarkdownMonster/_Classes/Commands/AppCommands.cs
@@ -479,6 +479,8 @@ namespace MarkdownMonster
                     }
                 }
 
+
+
                 var sd = new SaveFileDialog
                 {
                     FilterIndex = 1,