+++ MarkdownMonster/_Classes/Utilities/DocumentFileWatcher.cs
@@ -7,14 +7,26 @@ using MarkdownMonster.Windows;
 
 namespace MarkdownMonster.Utilities
 {
+
+    /// <summary>
+    /// File watcher used on the open editor document. Monitors the
+    /// file for changes and if a change occurs either updates the
+    /// document (if not dirty) or does nothing and defers to
+    /// editor.SaveDocument() to compare changes.
+    /// </summary>
     public class DocumentFileWatcher
     {
         private static FileSystemWatcher FileWatcher;
+
         private static readonly DebounceDispatcher debounce = new DebounceDispatcher();
 
-        public static void AttachFilewatcher(MarkdownDocumentEditor document)
+        /// <summary>
+        /// Attach an Editor instance to watch the related file
+        /// </summary>
+        /// <param name="editor">Editor instance on which to watch the file for changes</param>
+        public static void AttachFilewatcher(MarkdownDocumentEditor editor)
         {
-            var file = document?.MarkdownDocument?.Filename;
+            var file = editor?.MarkdownDocument?.Filename;
             if (string.IsNullOrEmpty(file) ||
                 file.Equals("untitled",StringComparison.InvariantCultureIgnoreCase) ||
                 !File.Exists(file))
@@ -50,65 +62,70 @@ namespace MarkdownMonster.Utilities
         /// <summary>
         /// If the file open in the active tab has changed try to
         /// either update the document if not dirty, otherwise
-        /// do nothing and wait for OnActivate to trigger a comparison.
+        /// do nothing and wait for save operation to trigger
+        /// comparison or choosing which file to pick (in SaveDocument())
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
         private static void FileWatcher_Changed(object sender, FileSystemEventArgs e)
         {
-            // Debug.WriteLine("FileWatcher: " + e.FullPath);
-
-            debounce.Debounce(200,(p) =>
+            debounce.Debounce(200,(objPath) =>
             {
-                // Debug.WriteLine("FileWatcher Debounce: " + e.FullPath);
+                string fullPath = objPath as string;
 
+                // Get the current editor and document
                 var tab = mmApp.Model.Window.TabControl.SelectedItem as TabItem;
                 var editor = tab?.Tag as MarkdownDocumentEditor;
                 var document = editor?.MarkdownDocument;
+                
+                bool isActiveTabFile = document != null && document.Filename.Equals(fullPath, StringComparison.InvariantCultureIgnoreCase);
+
+                if (!isActiveTabFile)
+                {
+                    tab = mmApp.Model.Window.GetTabFromFilename(fullPath);
+                    editor = tab?.Tag as MarkdownDocumentEditor;
+                    document = editor?.MarkdownDocument;
+                }
+
                 if (document == null)
                     return;
-                
+
+
+                // Only work changes on the active tab
                 // if the file was saved in the last second - don't update because
                 // the change event most likely is from the save op
                 // Note: this can include auto-save operations
                 if (document.LastSaveTime > DateTime.UtcNow.AddMilliseconds(-1000) ||
-                    document.IsDirty ||
-                    !document.Filename.Equals(e.FullPath,StringComparison.InvariantCultureIgnoreCase))
+                    document.IsDirty)
                     return;
-                
+              
+                // Otherwise reload from disk and refresh editor and preview
+
                 // load the document 
                 if (!document.Load(document.Filename))
                     return;
 
-                // Debug.WriteLine("FileWatcher Updating: " + e.FullPath);
-
-                // if the file was saved in the last second - don't update because
-                // the change event most likely is from the save op
-                // Note: this can include auto-save operations
-                if (document.LastSaveTime > DateTime.UtcNow.AddMilliseconds(-1000)) 
-                    return;
                 try
                 {
                     int scroll = editor.GetScrollPosition();
-                    editor.SetMarkdown(document.CurrentText);
-                    editor.AceEditor?.UpdateDocumentStats();
-
+                    editor.SetMarkdown(document.CurrentText, keepUndoBuffer: true);
+                    if (isActiveTabFile)
+                        editor.AceEditor?.UpdateDocumentStats();
                     if (scroll > -1)
                         editor.SetScrollPosition(scroll);
                 }
                 catch { }
 
-                if(document.ToString() != MarkdownDocument.PREVIEW_HTML_FILENAME)
+                if(isActiveTabFile &&
+                   document.ToString() != MarkdownDocument.PREVIEW_HTML_FILENAME)
                     mmApp.Model.Window.PreviewBrowser?.PreviewMarkdown(editor, keepScrollPosition: true);
 
-                //Debug.WriteLine("FileWatcher Updating DONE: " + e.FullPath);
-
-            },null, DispatcherPriority.Normal, mmApp.Model.Window.Dispatcher);
+            },e.FullPath, DispatcherPriority.Normal, mmApp.Model.Window.Dispatcher);
         }
 
 
         /// <summary>
-        /// Checks to see if 
+        /// Checks to see if a file has changed and updates the document in the editor
         /// </summary>
         /// <param name="noPrompt"></param>
         public static void CheckFileChangeInOpenDocuments(bool noPrompt = false)
@@ -132,20 +149,9 @@ namespace MarkdownMonster.Utilities
 
                 if (doc.HasFileCrcChanged())
                 {
-                    // force update to what's on disk so it doesn't fire again
-                    // do here prior to dialogs so this code doesn't fire recursively
-                    doc.UpdateCrc();
-
-                    string filename = doc.FilenamePathWithIndicator.Replace("*", "");
-                    string template = filename +
-                                      Environment.NewLine + Environment.NewLine +
-                                      "This file has been modified by another program." + Environment.NewLine +
-                                      "Do you want to reload it?";
-
-                    if (!doc.IsDirty || MessageBox.Show(window, template,
-                            "Reload",
-                            MessageBoxButton.YesNo,
-                            MessageBoxImage.Question, MessageBoxResult.No) == MessageBoxResult.Yes)
+                    // if the file is not dirty - reload. Otherwise it won't be notified until
+                    // you try and save the document.
+                    if (!doc.IsDirty)
                     {
                         if (!doc.Load(doc.Filename))
                         {